// src/file-upload/enums/file-extension.enum.ts
export enum FILE_EXTENSION {
  JSON = 'json',
  JS = 'js',
  COFFEE = 'coffee',
  HTML = 'html',
  TXT = 'txt',
  CSV = 'csv',
  PNG = 'png',
  JPG = 'jpg',
  JPEG = 'jpeg',
  GIF = 'gif',
  TIF = 'tif',
  TIFF = 'tiff',
  ZIP = 'zip',
  PDF = 'pdf',
  VCF = 'vcf',
  SVG = 'svg',
  XLS = 'xls',
  XLSX = 'xlsx',
  DOC = 'doc',
  DOCX = 'docx',
  MP3 = 'mp3',
}